libdpkg: Fix tar long name/link entry parsing to avoid truncation
A variable used during tar long name/link handling had an int type, but
was getting assigned an off_t value, which could then get truncated.

Change the variable from int to off_t to avoid this problem.

This is not a security issue as the tar code is only used for dpkg
operations which are not considered a security boundary.

Reported-by: Tristan <TristanInSec@gmail.com>

+++ lib/dpkg/tarfn.c
@@ -365,7 +365,7 @@ tar_gnu_long(struct tar_archive *tar, struct tar_entry *te, char **longp)
 	char buf[TARBLKSZ];
 	char *bp;
 	int status = 0;
-	int long_read;
+	off_t long_read;
 
 	free(*longp);
 	*longp = bp = m_malloc(te->size + 1);
